Refactor RouterEnsemblerConfig to use common schemas
